Moved db code to persist package and put it behind an interface
