+++ src/main/java/org/lumenframework/demo/notes/controllers/NoteViewController.java
@@ -9,20 +9,17 @@ import lumen.system.Annotations.Controller;
 import lumen.system.Annotations.Key;
 import lumen.system.Annotations.LumenEnabled;
 
-import org.lumenframework.demo.notes.DataStore;
-import org.lumenframework.demo.notes.Note;
+import org.lumenframework.demo.notes.persist.PersistableNoteH2Impl;
 
 import com.google.common.collect.Maps;
-import com.j256.ormlite.dao.Dao;
-import com.j256.ormlite.dao.DaoManager;
 
 @Controller
 public class NoteViewController {
 
     @LumenEnabled
     public static Component deleteNote(@Key("id")Long id, @Key("sort")String sort) throws Exception {
-        Dao<Note, Long> noteDao = DaoManager.createDao(DataStore.getInstance().getConnectionSource(), Note.class);
-        noteDao.deleteById(id);
+		PersistableNoteH2Impl persistImpl = new PersistableNoteH2Impl();
+		persistImpl.deleteById(id);
 
         Map<String, Object> listAttributes = Maps.newHashMap();
         listAttributes.put("sort", sort);
